bot/requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
environs==14.1.1
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
nats-py==2.10.0
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.19.0
